DEIMv2_Pose/requirements.txt
calflops==0.3.2
defusedxml==0.7.1
faster_coco_eval==1.7.0
ftfy==6.3.1
hydra-core==1.3.2
iopath==0.1.10
matplotlib==3.10.7
MultiScaleDeformableAttention==1.0
numpy==1.25.2
omegaconf==2.3.0
opencv_python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
pycocotools==2.0.10
PyYAML==6.0.3
regex==2024.11.6
rich==14.2.0
scikit_learn==1.7.2
scipy==1.16.3
setuptools==72.1.0
submitit==1.5.3
sympy==1.13.1
termcolor==3.2.0
torch==2.6.0
torchmetrics==1.8.2
torchvision==0.21.0
xformers==0.0.33.post1
